refactor: split the chain store implementation code
