+++ crates/amaru-stores/src/rocksdb/consensus/full_chain_store.rs
@@ -12,26 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod in_memory_chain_store;
-pub use in_memory_chain_store::*;
+use amaru_ouroboros_traits::FullChainStore;
 
-mod base_read_chain_store;
-pub use base_read_chain_store::*;
+use crate::rocksdb::consensus::RocksDBStore;
 
-mod read_chain_store;
-pub use read_chain_store::*;
-
-mod write_chain_store;
-pub use write_chain_store::*;
-
-mod types;
-pub use types::*;
-
-mod diagnostic_chain_store;
-pub use diagnostic_chain_store::*;
-
-mod full_chain_store;
-pub use full_chain_store::*;
-
-#[cfg(feature = "test-utils")]
-pub mod overriding_chain_store;
+impl FullChainStore for RocksDBStore {}
