+++ crates/amaru-stores/src/rocksdb/consensus/read_chain_store.rs
@@ -1,0 +1,39 @@
+// Copyright 2026 PRAGMA
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::fmt::Debug;
+
+use amaru_kernel::{IsHeader, cbor};
+use amaru_ouroboros_traits::{BaseReadChainStore, ReadChainStore};
+use rocksdb::{DB, OptimisticTransactionDB};
+
+use crate::rocksdb::consensus::RocksDBStore;
+
+impl<H> ReadChainStore<H> for RocksDBStore<OptimisticTransactionDB>
+where
+    H: IsHeader + Clone + Debug + for<'d> cbor::Decode<'d, ()>,
+{
+    fn snapshot(&self) -> Box<dyn BaseReadChainStore<H> + '_> {
+        Box::new(RocksDBStore { basedir: self.basedir.clone(), db: self.db.snapshot() })
+    }
+}
+
+impl<H> ReadChainStore<H> for RocksDBStore<DB>
+where
+    H: IsHeader + Clone + Debug + for<'d> cbor::Decode<'d, ()>,
+{
+    fn snapshot(&self) -> Box<dyn BaseReadChainStore<H> + '_> {
+        Box::new(RocksDBStore { basedir: self.basedir.clone(), db: self.db.snapshot() })
+    }
+}
